requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0
redis==5.1.0
arq==0.26.1
anthropic==0.40.0
pydantic-settings==2.5.2
pydantic==2.9.2
sse-starlette==2.1.3
httpx==0.27.2
pyyaml==6.0.2
jinja2==3.1.4
pytest==8.3.3
pytest-asyncio==0.24.0
anyio==4.6.2
fakeredis==2.26.1
respx==0.21.1
ruff==0.6.9
mypy==1.11.2
pytest-cov==5.0.0
